+++ jquery.caption.js
@@ -6,8 +6,8 @@
  * Released under the MIT license
  * https://github.com/jquery/jquery/blob/master/MIT-LICENSE.txt
  *
- * v0.9.6
- * Date: 2014-09-03
+ * v0.9.7
+ * Date: 2014-10-13
  */
 (function($, window, undefined){
 	$.fn.captionjs = function(opts){
@@ -16,14 +16,24 @@
 		var defaults = {
 			'class_name'      : 'captionjs', // Class name assigned to each <figure>
 			'schema'          : true,        // Use schema.org markup (i.e., itemtype, itemprop)
-			'mode'            : 'default',   // default | static | animated | hide
+			'mode'            : 'default',   // default | stacked | animated | hide
 			'debug_mode'      : false,       // Output debug info to the JS console
-			'force_dimensions': false,       // Force the dimensions in case they can't be detected (e.g., image is not yet painted to viewport)
-			'is_responsive'   : false        // Ensure the figure and image change size when in responsive layout. Requires a container to control responsiveness!
+			'force_dimensions': true,        // Force the dimensions in case they can't be detected (e.g., image is not yet painted to viewport)
+			'is_responsive'   : false,       // Ensure the figure and image change size when in responsive layout. Requires a container to control responsiveness!
+			'inherit_styles'  : true         // Have the caption.js container inherit all box-model styles from the original image
 		};
 
 		// Extend the options from defaults with user's options
 		var options = $.extend(defaults, opts || {});
+
+		// Function to copy styles
+		var transferStyles = function(property, reset_val, $origin, $target){
+			if ($origin.jquery && $target.jquery) // Check that they are jQuery objects
+			{
+				$origin.css(property, $target.css(property));
+				$target.css(property, reset_val);
+			}
+		};
 
 		// jQuery chainability -- do the magic below
 		return this.each(function(){
@@ -48,15 +58,14 @@
 			{
 				if (options.debug_mode) console.log('caption.js | Forcing dimensions with a clone.');
 
-				// Create a clone outside of the viewport to detect and
-				// then set the dimensions.
+				// Create a clone outside of the viewport to detect and then set the dimensions.
 				var $clone = $figure.clone().css({
 					'position': 'absolute',
 					'left'    : '-9999px'
 				}).appendTo('body');
 
 				target_width = $('img', $clone).outerWidth();
-				target_height = $('figcaption', $clone).css('width', target_width).outerHeight(); // Make sure width (and thus line wrapping) is enforced so that the height is correct
+				target_height = $('figcaption', $clone).css('width', target_width).css('clear', 'both').outerHeight(); // Make sure width (and thus line wrapping) is enforced so that the height is correct
 
 				$clone.remove();
 			}
@@ -72,6 +81,31 @@
 			{
 				target_width = '100%';
 				$this.width(target_width);
+			}
+
+			// Copy styles if need be
+			if (options.inherit_styles)
+			{
+				if ($this.css('display') == 'inline')
+					$figure.css('display', 'inline-block');
+				else
+					transferStyles('display', 'block', $figure, $this);
+
+				if ($this.css('position') == 'static')
+					$figure.css('position', 'relative');
+				else
+					transferStyles('position', 'relative', $figure, $this);
+
+				transferStyles('clear', 'both', $figure, $this);
+				transferStyles('float', 'none', $figure, $this);
+				transferStyles('margin', '0', $figure, $this);
+				// transferStyles('padding', '0', $figure, $this);
+				$this.css('padding', '0');
+				transferStyles('left', 'auto', $figure, $this);
+				transferStyles('right', 'auto', $figure, $this);
+				transferStyles('top', 'auto', $figure, $this);
+				transferStyles('bottom', 'auto', $figure, $this);
+				transferStyles('z-index', $this.css('z-index'), $figure, $this);
 			}
 
 			// Set the width of the figure.